Split all code into domain ui and api folders
