+++ src/domain/dndHelpers.ts
