+++ src/domain/operations.ts
