+++ src/domain/state.spec.ts
